pyproject.toml
[project]
name = "langgraph-agentic-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ddgs>=9.10.0",
    "faiss-cpu>=1.13.2",
    "ipython>=9.9.0",
    "langchain>=1.2.7",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.7",
    "pypdf>=6.6.2",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.1",
    "tavily-python>=0.7.19",
]


